refactor(api): migrate shareController to ts-rest + split into folder
Complete the ts-rest migration of /api/share and break up the 1600-line
legacy shareController:

- Remove the 6 dead duplicate handlers that the existing shareContractRouter
  (mounted first) already shadowed (image/video/video-from-project/update-
  image/save-as-template/push-to-phone).
- New sharesReadContract + shareReadContractRouter.ts for the auth-guarded
  read/management routes: publish, my, recent, my/images, my/videos, delete,
  templates (clone/list/get), devices. Auth enforced per handler (the prefix
  uses optionalAuth). Route keys ordered so no /:shareToken catch-all is
  introduced in the contract.
- shareFileRouter.ts keeps the public info + file-streaming routes on legacy
  Express (range streaming, dynamic thumbnails, two-stage download) under the
  existing publicReadLimiter.
- shareServices.ts holds the shared lazy service factory + background render.
- Extract toCamelCase to apps/api/utils/case.ts (+ vitest).
- routes.ts mounts: write contract → read contract → legacy file router.

Response validation stays off (compile-time schemas only). No behavior change.
