--- apps/api/routes/share/shareController.ts
@@ -1,1600 +1,0 @@
-/**
- * Share Controller
- * Handles sharing of images and videos with public links
- */
-
-import crypto from 'crypto';
-import fs from 'fs';
-import path from 'path';
-
-import express, { type Request, type Response, type Router } from 'express';
-import rateLimit from 'express-rate-limit';
-import sharp from 'sharp';
-import { z } from 'zod';
-
-import { requireAuth } from '../../middleware/authMiddleware.js';
-import { validateBody, type TypedRequest } from '../../middleware/validateBody.js';
-import { transferService } from '../../services/transferService.js';
-import { type SharedMediaRow, type ShareResult } from '../../types/media.js';
-import { setContentDisposition } from '../../utils/http/contentDisposition.js';
-import { createLogger } from '../../utils/logger.js';
-import { parseJSON } from '../../utils/parseJSON.js';
-import { redisClient } from '../../utils/redis/index.js';
-
-import type { AuthenticatedRequest } from '../../middleware/types.js';
-
-const fsPromises = fs.promises;
-const log = createLogger('share');
-
-// Rate limiters for share mutation routes
-const shareWriteLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 50,
-  standardHeaders: true,
-  legacyHeaders: false,
-});
-
-// ============================================================================
-// Types
-// ============================================================================
-
-interface CamelCaseObject {
-  [key: string]: unknown;
-}
-
-interface ExportData {
-  status: string;
-  outputPath: string;
-  duration?: number;
-}
-
-interface Project {
-  id?: string;
-  video_path?: string;
-  thumbnail_path?: string;
-  subtitled_video_path?: string;
-  subtitles?: unknown;
-  title?: string;
-  video_metadata?: {
-    width?: number;
-    height?: number;
-    duration?: number;
-  };
-  style_preference?: string;
-  height_preference?: string;
-}
-
-interface SharedMediaService {
-  ensureInitialized(): Promise<void>;
-  createImageShare(userId: string, params: CreateImageShareParams): Promise<ShareResult>;
-  createVideoShare(userId: string, params: CreateVideoShareParams): Promise<ShareResult>;
-  createPendingVideoShare(
-    userId: string,
-    params: CreatePendingVideoShareParams
-  ): Promise<ShareResult>;
-  getUserShares(
-    userId: string,
-    type: string | null,
-    status?: string | null
-  ): Promise<SharedMediaRow[]>;
-  getUserShareCount(userId: string): Promise<number>;
-  getShareByToken(shareToken: string): Promise<SharedMediaRow | null>;
-  recordView(shareToken: string): Promise<void>;
-  recordDownload(
-    shareToken: string,
-    email: string | null,
-    ip: string,
-    shareId?: string
-  ): Promise<void>;
-  deleteShare(userId: string, shareToken: string): Promise<void>;
-  finalizeVideoShare(shareToken: string, videoPath: string): Promise<void>;
-  markShareFailed(shareToken: string): Promise<void>;
-  updateImageShare(
-    userId: string,
-    shareToken: string,
-    params: UpdateImageShareParams
-  ): Promise<ShareResult>;
-  getThumbnailFilePath(relativePath: string): string;
-  getMediaFilePath(relativePath: string): string;
-  getOriginalImagePath(shareToken: string, filename: string): string;
-  clearOriginalImageMetadata(shareToken: string): Promise<void>;
-  markAsTemplate(
-    userId: string,
-    shareToken: string,
-    title: string,
-    visibility: string,
-    userName: string
-  ): Promise<void>;
-  cloneTemplate(templateToken: string, userId: string, userName: string): Promise<ShareResult>;
-  getTemplates(userId: string | null, visibility: string): Promise<SharedMediaRow[]>;
-  getTemplateByToken(
-    templateToken: string,
-    requestingUserId?: string
-  ): Promise<SharedMediaRow | null>;
-}
-
-interface ProjectService {
-  ensureInitialized(): Promise<void>;
-  getProject(userId: string, projectId: string): Promise<Project>;
-  getVideoPath(relativePath: string): string;
-  getSubtitledVideoPath(relativePath: string): string;
-  getThumbnailPath(relativePath: string): string;
-  updateSubtitledVideoPath(userId: string, projectId: string, relativePath: string): Promise<void>;
-}
-
-interface CreateImageShareParams {
-  imageBase64: string;
-  title: string;
-  imageType: string | null;
-  metadata: Record<string, unknown>;
-  originalImage: string | null;
-  status?: 'ready' | 'draft';
-}
-
-interface CreateVideoShareParams {
-  videoPath: string;
-  title: string;
-  thumbnailPath: string | null;
-  duration: number | null;
-  projectId: string | null;
-}
-
-interface CreatePendingVideoShareParams {
-  title: string;
-  thumbnailPath: string | null;
-  duration: number | null;
-  projectId: string;
-}
-
-interface UpdateImageShareParams {
-  imageBase64: string;
-  title?: string;
-  metadata: Record<string, unknown>;
-  originalImage?: string | null;
-}
-
-interface ShareTokenParams {
-  shareToken: string;
-  [key: string]: string;
-}
-
-/** Request extended with pre-fetched share data passed between middleware handlers */
-interface RequestWithShare extends Request<ShareTokenParams> {
-  _share?: SharedMediaRow;
-}
-
-interface ShareResponse {
-  success: boolean;
-  share?: {
-    shareToken: string;
-    shareUrl: string;
-    createdAt: Date | string;
-    mediaType: 'image' | 'video';
-    hasOriginalImage?: boolean | undefined;
-    status?: string | undefined;
-  };
-  error?: string;
-  code?: string;
-}
-
-interface ShareListResponse {
-  success: boolean;
-  shares?: CamelCaseObject[];
-  count?: number;
-  limit?: number;
-  error?: string;
-}
-
-interface ShareInfoResponse {
-  success: boolean;
-  share?: {
-    mediaType: string;
-    title: string | null;
-    thumbnailUrl: string | null;
-    downloadCount: number;
-    viewCount: number;
-    sharerName?: string;
-    status: string;
-    createdAt: Date;
-    fileName?: string | null;
-    fileSize?: number | null;
-    mimeType?: string | null;
-    isPasswordProtected?: boolean;
-    expiresAt?: Date | null;
-    transferMessage?: string | null;
-    transferFiles?: Array<{ name: string; size: number; mimeType: string }> | null;
-    duration?: number | null;
-    imageType?: string | null;
-    dimensions?: {
-      width?: number | undefined;
-      height?: number | undefined;
-    };
-  };
-  error?: string;
-}
-
-// ============================================================================
-// Helper Functions
-// ============================================================================
-
-function toCamelCase(obj: unknown): unknown {
-  if (Array.isArray(obj)) {
-    return obj.map(toCamelCase);
-  }
-  if (obj !== null && typeof obj === 'object') {
-    return Object.entries(obj as Record<string, unknown>).reduce<CamelCaseObject>(
-      (acc, [key, value]) => {
-        const camelKey = key.replace(/_([a-z])/g, (_, letter: string) => letter.toUpperCase());
-        acc[camelKey] = toCamelCase(value);
-        return acc;
-      },
-      {}
-    );
-  }
-  return obj;
-}
-
-// Lazy-loaded services
-let sharedMediaService: SharedMediaService | null = null;
-let projectService: ProjectService | null = null;
-
-async function getSharedMediaService(): Promise<SharedMediaService> {
-  if (!sharedMediaService) {
-    const { getSharedMediaService: getService } =
-      await import('../../services/sharedMediaService.js');
-    sharedMediaService = getService() as unknown as SharedMediaService;
-    await sharedMediaService.ensureInitialized();
-  }
-  return sharedMediaService;
-}
-
-async function getProjectService(): Promise<ProjectService> {
-  if (!projectService) {
-    const { getSubtitlerProjectService } = await import('../../services/subtitler/index.js');
-    projectService = getSubtitlerProjectService() as unknown as ProjectService;
-    await projectService.ensureInitialized();
-  }
-  return projectService;
-}
-
-async function triggerBackgroundRender(
-  userId: string,
-  projectId: string,
-  shareToken: string,
-  project: Project
-): Promise<void> {
-  try {
-    const projService = await getProjectService();
-    const { processProjectExport } = await import('../../services/subtitler/exportService.js');
-
-    log.info(`Background render starting for share ${shareToken}`);
-
-    const result = await processProjectExport(
-      project as {
-        id: string;
-        video_path: string;
-        subtitles: string;
-        style_preference?: string;
-        height_preference?: string;
-      },
-      projService
-    );
-
-    const subtitledVideoRelativePath = `${userId}/${projectId}/subtitled_${Date.now()}.mp4`;
-    const subtitledVideoFullPath = projService.getSubtitledVideoPath(subtitledVideoRelativePath);
-
-    await fsPromises.mkdir(path.dirname(subtitledVideoFullPath), { recursive: true });
-    await fsPromises.copyFile(result.outputPath, subtitledVideoFullPath);
-    await projService.updateSubtitledVideoPath(userId, projectId, subtitledVideoRelativePath);
-
-    const service = await getSharedMediaService();
-    await service.finalizeVideoShare(shareToken, subtitledVideoFullPath);
-
-    try {
-      await fsPromises.unlink(result.outputPath);
-    } catch {
-      // Ignore cleanup errors
-    }
-
-    log.info(`Background render complete for share ${shareToken}`);
-  } catch (error) {
-    log.error(`Background render failed for ${shareToken}:`, error);
-    const service = await getSharedMediaService();
-    await service.markShareFailed(shareToken);
-  }
-}
-
-// ============================================================================
-// Zod Schemas
-// ============================================================================
-
-const createImageShareSchema = z.object({
-  imageData: z.string(),
-  title: z.string().optional(),
-  imageType: z.string().optional(),
-  metadata: z.record(z.unknown()).optional(),
-  originalImage: z.string().optional(),
-  status: z.enum(['ready', 'draft']).optional(),
-});
-type CreateImageShareBody = z.infer<typeof createImageShareSchema>;
-
-const createVideoShareSchema = z.object({
-  exportToken: z.string(),
-  title: z.string().optional(),
-  projectId: z.string().optional(),
-});
-type CreateVideoShareBody = z.infer<typeof createVideoShareSchema>;
-
-const createVideoFromProjectSchema = z.object({
-  projectId: z.string(),
-  title: z.string().optional(),
-});
-type CreateVideoFromProjectBody = z.infer<typeof createVideoFromProjectSchema>;
-
-const saveAsTemplateSchema = z.object({
-  title: z.string().optional(),
-  visibility: z.enum(['private', 'unlisted', 'public']).optional(),
-});
-type SaveAsTemplateBody = z.infer<typeof saveAsTemplateSchema>;
-
-const updateImageShareSchema = z.object({
-  imageBase64: z.string(),
-  title: z.string().optional(),
-  metadata: z.record(z.unknown()).optional(),
-  originalImage: z.string().optional(),
-});
-type UpdateImageShareBody = z.infer<typeof updateImageShareSchema>;
-
-const pushToPhoneSchema = z.object({
-  shareToken: z.string(),
-});
-type PushToPhoneBody = z.infer<typeof pushToPhoneSchema>;
-
-// ============================================================================
-// Router Setup
-// ============================================================================
-
-const router: Router = express.Router();
-
-// Override global body parser limit for share routes (canvas images can exceed 10MB)
-router.use(express.json({ limit: '50mb' }));
-
-// ============================================================================
-// IMAGE SHARE ROUTES
-// ============================================================================
-
-router.post(
-  '/image',
-  shareWriteLimiter,
-  requireAuth,
-  validateBody(createImageShareSchema),
-  async (req: TypedRequest<CreateImageShareBody>, res: Response<ShareResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const { imageData, title, imageType, metadata, originalImage, status } = req.body;
-
-      const service = await getSharedMediaService();
-      const share = await service.createImageShare(userId, {
-        imageBase64: imageData,
-        title: title || 'Geteiltes Bild',
-        imageType: imageType || null,
-        metadata: metadata || {},
-        originalImage: originalImage || null,
-        status: status === 'draft' ? 'draft' : 'ready',
-      });
-
-      log.info(
-        `Image share created: ${share.shareToken} by user ${userId}${originalImage ? ' (with original)' : ''}`
-      );
-
-      return res.json({
-        success: true,
-        share: {
-          shareToken: share.shareToken,
-          shareUrl: share.shareUrl,
-          createdAt: share.createdAt,
-          mediaType: 'image',
-          hasOriginalImage: share.hasOriginalImage || false,
-        },
-      });
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      const isValidation = message.startsWith('createImageShare:');
-      log.error('Failed to create image share:', error);
-      return res.status(isValidation ? 400 : 500).json({
-        success: false,
-        error: isValidation ? message : 'Bild konnte nicht geteilt werden',
-      });
-    }
-  }
-);
-
-// Promote a draft to ready (publish)
-router.put(
-  '/:shareToken/publish',
-  shareWriteLimiter,
-  requireAuth,
-  async (req: Request<ShareTokenParams>, res: Response<ShareResponse>) => {
-    try {
-      const userId = (req as AuthenticatedRequest).user!.id;
-      const { shareToken } = req.params;
-
-      const service = await getSharedMediaService();
-      const share = await service.getShareByToken(shareToken as string);
-
-      if (!share) {
-        return res.status(404).json({ success: false, error: 'Share nicht gefunden' });
-      }
-
-      if (share.user_id !== userId) {
-        return res.status(403).json({ success: false, error: 'Nicht berechtigt' });
-      }
-
-      const pg = (await import('../../database/services/PostgresService.js')).getPostgresInstance();
-      await pg.query('UPDATE shared_media SET status = $1 WHERE share_token = $2', [
-        'ready',
-        shareToken,
-      ]);
-
-      log.info(`Share ${shareToken} published by user ${userId}`);
-
-      return res.json({
-        success: true,
-        share: {
-          shareToken: share.share_token,
-          shareUrl: `/share/${shareToken}`,
-          createdAt: share.created_at,
-          mediaType: share.media_type as 'image' | 'video',
-          status: 'ready',
-        },
-      });
-    } catch (error) {
-      log.error('Failed to publish share:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Share konnte nicht veröffentlicht werden',
-      });
-    }
-  }
-);
-
-// ============================================================================
-// VIDEO SHARE ROUTES
-// ============================================================================
-
-router.post(
-  '/video',
-  requireAuth,
-  validateBody(createVideoShareSchema),
-  async (req: TypedRequest<CreateVideoShareBody>, res: Response<ShareResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const { exportToken, title, projectId } = req.body;
-
-      const exportDataString = (await redisClient.get(`export:${exportToken}`)) as string | null;
-      if (!exportDataString) {
-        return res.status(404).json({
-          success: false,
-          error: 'Export nicht gefunden oder abgelaufen',
-        });
-      }
-
-      const exportData = parseJSON<ExportData>(exportDataString);
-      if (exportData.status !== 'complete') {
-        return res.status(400).json({
-          success: false,
-          error: 'Export noch nicht abgeschlossen',
-        });
-      }
-
-      const { outputPath, duration } = exportData;
-
-      try {
-        await fsPromises.access(outputPath);
-      } catch {
-        return res.status(404).json({
-          success: false,
-          error: 'Export-Datei nicht gefunden',
-        });
-      }
-
-      const service = await getSharedMediaService();
-      const share = await service.createVideoShare(userId, {
-        videoPath: outputPath,
-        title: title || 'Geteiltes Video',
-        thumbnailPath: null,
-        duration: duration || null,
-        projectId: projectId || null,
-      });
-
-      log.info(`Video share created: ${share.shareToken} by user ${userId}`);
-
-      return res.json({
-        success: true,
-        share: {
-          shareToken: share.shareToken,
-          shareUrl: share.shareUrl,
-          createdAt: share.createdAt,
-          mediaType: 'video',
-          status: 'ready',
-        },
-      });
-    } catch (error) {
-      log.error('Failed to create video share:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Video konnte nicht geteilt werden',
-      });
-    }
-  }
-);
-
-router.post(
-  '/video/from-project',
-  requireAuth,
-  validateBody(createVideoFromProjectSchema),
-  async (req: TypedRequest<CreateVideoFromProjectBody>, res: Response<ShareResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const { projectId, title } = req.body;
-
-      const projService = await getProjectService();
-      let project: Project;
-      try {
-        project = await projService.getProject(userId, projectId);
-      } catch {
-        return res.status(404).json({
-          success: false,
-          error: 'Projekt nicht gefunden oder keine Berechtigung',
-        });
-      }
-
-      if (!project || !project.video_path) {
-        return res.status(404).json({
-          success: false,
-          error: 'Projekt-Video nicht gefunden',
-        });
-      }
-
-      let thumbnailPath: string | null = null;
-      if (project.thumbnail_path) {
-        thumbnailPath = projService.getThumbnailPath(project.thumbnail_path);
-        try {
-          await fsPromises.access(thumbnailPath);
-        } catch {
-          thumbnailPath = null;
-        }
-      }
-
-      const service = await getSharedMediaService();
-
-      if (!project.subtitled_video_path) {
-        if (!project.subtitles) {
-          return res.status(400).json({
-            success: false,
-            error: 'Projekt hat keine Untertitel zum Exportieren.',
-            code: 'NO_SUBTITLES',
-          });
-        }
-
-        const share = await service.createPendingVideoShare(userId, {
-          title: title || project.title || 'Geteiltes Video',
-          thumbnailPath: thumbnailPath,
-          duration: project.video_metadata?.duration || null,
-          projectId: projectId,
-        });
-
-        void triggerBackgroundRender(userId, projectId, share.shareToken, project);
-
-        log.info(
-          `Video share created (rendering): ${share.shareToken} for project ${projectId} by user ${userId}`
-        );
-
-        return res.json({
-          success: true,
-          share: {
-            shareToken: share.shareToken,
-            shareUrl: share.shareUrl,
-            createdAt: share.createdAt,
-            mediaType: 'video',
-            status: 'processing',
-          },
-        });
-      }
-
-      const videoPath = projService.getSubtitledVideoPath(project.subtitled_video_path);
-
-      try {
-        await fsPromises.access(videoPath);
-      } catch {
-        if (!project.subtitles) {
-          return res.status(400).json({
-            success: false,
-            error: 'Video-Datei nicht gefunden und keine Untertitel zum Rendern.',
-            code: 'NO_SUBTITLES',
-          });
-        }
-
-        const share = await service.createPendingVideoShare(userId, {
-          title: title || project.title || 'Geteiltes Video',
-          thumbnailPath: thumbnailPath,
-          duration: project.video_metadata?.duration || null,
-          projectId: projectId,
-        });
-
-        void triggerBackgroundRender(userId, projectId, share.shareToken, project);
-
-        log.info(
-          `Video share created (re-rendering): ${share.shareToken} for project ${projectId} by user ${userId}`
-        );
-
-        return res.json({
-          success: true,
-          share: {
-            shareToken: share.shareToken,
-            shareUrl: share.shareUrl,
-            createdAt: share.createdAt,
-            mediaType: 'video',
-            status: 'processing',
-          },
-        });
-      }
-
-      const share = await service.createVideoShare(userId, {
-        videoPath: videoPath,
-        title: title || project.title || 'Geteiltes Video',
-        thumbnailPath: thumbnailPath,
-        duration: project.video_metadata?.duration || null,
-        projectId: projectId,
-      });
-
-      log.info(
-        `Video share created from project: ${share.shareToken} for project ${projectId} by user ${userId}`
-      );
-
-      return res.json({
-        success: true,
-        share: {
-          shareToken: share.shareToken,
-          shareUrl: share.shareUrl,
-          createdAt: share.createdAt,
-          mediaType: 'video',
-          status: 'ready',
-        },
-      });
-    } catch (error) {
-      log.error('Failed to create video share from project:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Video konnte nicht geteilt werden',
-      });
-    }
-  }
-);
-
-// ============================================================================
-// UNIFIED ROUTES
-// ============================================================================
-
-router.get(
-  '/my',
-  requireAuth,
-  async (req: AuthenticatedRequest, res: Response<ShareListResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const type = req.query.type as string | undefined;
-      const status = req.query.status as string | undefined;
-
-      const service = await getSharedMediaService();
-      const shares = await service.getUserShares(userId, type || null, status || null);
-      const count = await service.getUserShareCount(userId);
-
-      res.json({
-        success: true,
-        shares: toCamelCase(shares) as CamelCaseObject[],
-        count,
-        limit: 50,
-      });
-    } catch (error) {
-      log.error('Failed to get user shares:', error);
-      res.status(500).json({
-        success: false,
-        error: 'Geteilte Medien konnten nicht geladen werden',
-      });
-    }
-  }
-);
-
-router.get(
-  '/recent',
-  requireAuth,
-  async (req: AuthenticatedRequest, res: Response<ShareListResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const limit = Math.min(parseInt(req.query.limit as string) || 6, 20);
-
-      let service;
-      try {
-        service = await getSharedMediaService();
-      } catch (initError) {
-        log.warn('SharedMediaService unavailable, returning empty result:', initError);
-        return res.json({
-          success: true,
-          shares: [],
-          count: 0,
-          limit,
-        });
-      }
-
-      const allShares = await service.getUserShares(userId, 'image');
-      const recentShares = allShares.slice(0, limit);
-
-      res.json({
-        success: true,
-        shares: toCamelCase(recentShares) as CamelCaseObject[],
-        count: recentShares.length,
-        limit,
-      });
-    } catch (error) {
-      log.error('Failed to get recent shares:', error);
-      res.status(500).json({
-        success: false,
-        error: 'Letzte Bilder konnten nicht geladen werden',
-      });
-    }
-  }
-);
-
-router.get(
-  '/my/images',
-  requireAuth,
-  async (req: AuthenticatedRequest, res: Response<ShareListResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const service = await getSharedMediaService();
-      const shares = await service.getUserShares(userId, 'image');
-
-      res.json({
-        success: true,
-        shares: toCamelCase(shares) as CamelCaseObject[],
-      });
-    } catch (error) {
-      log.error('Failed to get user image shares:', error);
-      res.status(500).json({
-        success: false,
-        error: 'Bilder konnten nicht geladen werden',
-      });
-    }
-  }
-);
-
-router.get(
-  '/my/videos',
-  requireAuth,
-  async (req: AuthenticatedRequest, res: Response<ShareListResponse>) => {
-    try {
-      const userId = req.user!.id;
-      const service = await getSharedMediaService();
-      const shares = await service.getUserShares(userId, 'video');
-
-      res.json({
-        success: true,
-        shares: toCamelCase(shares) as CamelCaseObject[],
-      });
-    } catch (error) {
-      log.error('Failed to get user video shares:', error);
-      res.status(500).json({
-        success: false,
-        error: 'Videos konnten nicht geladen werden',
-      });
-    }
-  }
-);
-
-router.get(
-  '/:shareToken',
-  async (req: Request<ShareTokenParams>, res: Response<ShareInfoResponse>) => {
-    try {
-      const { shareToken } = req.params;
-      const service = await getSharedMediaService();
-      const share = await service.getShareByToken(shareToken);
-
-      if (!share) {
-        return res.status(404).json({
-          success: false,
-          error: 'Geteiltes Medium nicht gefunden',
-        });
-      }
-
-      await service.recordView(shareToken);
-
-      const shareObj: NonNullable<ShareInfoResponse['share']> = {
-        mediaType: share.media_type,
-        title: share.title,
-        thumbnailUrl: share.thumbnail_path ? `/api/share/${shareToken}/thumbnail` : null,
-        downloadCount: share.download_count,
-        viewCount: share.view_count,
-        status: share.status || 'ready',
-        createdAt: share.created_at,
-        ...(share.sharer_name != null ? { sharerName: share.sharer_name } : {}),
-      };
-      const response: ShareInfoResponse = {
-        success: true,
-        share: shareObj,
-      };
-
-      if (share.media_type === 'transfer') {
-        // Check expiry for transfers
-        if (share.expires_at && new Date(share.expires_at) < new Date()) {
-          return res.status(410).json({
-            success: false,
-            error: 'Dieser Transfer-Link ist abgelaufen.',
-          });
-        }
-
-        response.share!.fileName = share.file_name;
-        response.share!.fileSize = share.file_size;
-        response.share!.mimeType = share.mime_type;
-        response.share!.isPasswordProtected = !!share.password_hash;
-        response.share!.expiresAt = share.expires_at ?? null;
-        response.share!.transferMessage = share.transfer_message ?? null;
-
-        const files = share.transfer_files;
-        if (Array.isArray(files) && files.length > 0) {
-          response.share!.transferFiles = files.map(
-            (f: { name: string; size: number; mimeType: string }) => ({
-              name: f.name,
-              size: f.size,
-              mimeType: f.mimeType,
-            })
-          );
-        }
-      } else if (share.media_type === 'video') {
-        response.share!.duration = share.duration;
-      } else {
-        const metadata = (
-          typeof share.image_metadata === 'string'
-            ? JSON.parse(share.image_metadata)
-            : share.image_metadata || {}
-        ) as { width?: number; height?: number };
-        response.share!.imageType = share.image_type;
-        response.share!.dimensions = {
-          width: metadata.width,
-          height: metadata.height,
-        };
-      }
-
-      return res.json(response);
-    } catch (error) {
-      log.error('Failed to get share info:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Fehler beim Laden des geteilten Mediums',
-      });
-    }
-  }
-);
-
-router.get('/:shareToken/thumbnail', async (req: Request<ShareTokenParams>, res: Response) => {
-  try {
-    const { shareToken } = req.params;
-    const service = await getSharedMediaService();
-    const share = await service.getShareByToken(shareToken);
-
-    if (!share || !share.thumbnail_path) {
-      return res.status(404).json({ error: 'Thumbnail nicht gefunden' });
-    }
-
-    const thumbnailPath = service.getThumbnailFilePath(share.thumbnail_path);
-
-    try {
-      await fsPromises.access(thumbnailPath);
-      return res.sendFile(thumbnailPath);
-    } catch {
-      return res.status(404).json({ error: 'Thumbnail-Datei nicht gefunden' });
-    }
-  } catch (error) {
-    log.error('Failed to get thumbnail:', error);
-    return res.status(500).json({ error: 'Fehler beim Laden des Thumbnails' });
-  }
-});
-
-router.get(
-  '/:shareToken/original',
-  requireAuth,
-  async (req: Request<ShareTokenParams> & AuthenticatedRequest, res: Response) => {
-    try {
-      const { shareToken } = req.params;
-      const userId = req.user!.id;
-
-      const service = await getSharedMediaService();
-      const share = await service.getShareByToken(shareToken);
-
-      if (!share) {
-        return res.status(404).json({ error: 'Share nicht gefunden' });
-      }
-
-      if (share.user_id !== userId) {
-        return res.status(403).json({ error: 'Zugriff verweigert' });
-      }
-
-      const metadata = (share.image_metadata || {}) as Record<string, unknown>;
-      if (!metadata.hasOriginalImage || !metadata.originalImageFilename) {
-        return res.status(404).json({ error: 'Originalbild nicht vorhanden' });
-      }
-
-      const originalPath = service.getOriginalImagePath(
-        shareToken,
-        metadata.originalImageFilename as string
-      );
-
-      try {
-        const stat = await fsPromises.stat(originalPath);
-        const mimeType = (metadata.originalImageFilename as string).endsWith('.jpg')
-          ? 'image/jpeg'
-          : 'image/png';
-
-        res.setHeader('Content-Type', mimeType);
-        res.setHeader('Content-Length', stat.size);
-        res.setHeader('Cache-Control', 'private, max-age=3600');
-
-        return fs.createReadStream(originalPath).pipe(res);
-      } catch {
-        // File is missing on disk but metadata claims it exists. Self-heal so
-        // the gallery edit handler stops generating /original URLs for this share.
-        try {
-          await service.clearOriginalImageMetadata(shareToken);
-        } catch (healError) {
-          log.warn(
-            'Failed to clear stale originalImage metadata for %s: %s',
-            shareToken,
-            healError
-          );
-        }
-        return res.status(404).json({ error: 'Originalbild-Datei nicht gefunden' });
-      }
-    } catch (error) {
-      log.error('Failed to get original image:', error);
-      return res.status(500).json({ error: 'Fehler beim Laden des Originalbildes' });
-    }
-  }
-);
-
-router.put(
-  '/:shareToken/image',
-  requireAuth,
-  validateBody(updateImageShareSchema),
-  async (
-    req: TypedRequest<UpdateImageShareBody, ShareTokenParams>,
-    res: Response<ShareResponse>
-  ) => {
-    try {
-      const { shareToken } = req.params;
-      const userId = req.user!.id;
-      const { imageBase64, title, metadata, originalImage } = req.body;
-
-      const service = await getSharedMediaService();
-
-      const existingShare = await service.getShareByToken(shareToken);
-      if (!existingShare) {
-        return res.status(404).json({
-          success: false,
-          error: 'Share nicht gefunden',
-        });
-      }
-
-      if (existingShare.user_id !== userId) {
-        return res.status(403).json({
-          success: false,
-          error: 'Nur die erstellende Person kann diesen Share bearbeiten',
-        });
-      }
-
-      if (existingShare.media_type !== 'image') {
-        return res.status(400).json({
-          success: false,
-          error: 'Nur Bild-Shares können aktualisiert werden',
-        });
-      }
-
-      const updateParams: UpdateImageShareParams = {
-        imageBase64,
-        metadata: metadata || {},
-        ...(title != null ? { title } : {}),
-        ...(originalImage != null ? { originalImage } : {}),
-      };
-      const result = await service.updateImageShare(userId, shareToken, updateParams);
-
-      log.info(`Image share updated: ${shareToken} by user ${userId}`);
-
-      const shareResp: ShareResult = {
-        id: result.id,
-        shareToken: result.shareToken,
-        shareUrl: result.shareUrl,
-        createdAt: result.createdAt,
-        mediaType: 'image',
-        ...(result.hasOriginalImage != null ? { hasOriginalImage: result.hasOriginalImage } : {}),
-      };
-      return res.json({
-        success: true,
-        share: shareResp,
-      });
-    } catch (error) {
-      log.error('Failed to update image share:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Bild konnte nicht aktualisiert werden',
-      });
-    }
-  }
-);
-
-router.get('/:shareToken/preview', async (req: Request<ShareTokenParams>, res: Response) => {
-  try {
-    const { shareToken } = req.params;
-    const service = await getSharedMediaService();
-    const share = await service.getShareByToken(shareToken);
-
-    if (!share) {
-      return res.status(404).json({ error: 'Medium nicht gefunden' });
-    }
-
-    if (share.status === 'processing') {
-      return res.status(202).json({
-        status: 'processing',
-        message: 'Medium wird noch verarbeitet',
-      });
-    }
-
-    if (share.status === 'failed') {
-      return res.status(500).json({ error: 'Verarbeitung fehlgeschlagen' });
-    }
-
-    if (!share.file_path) {
-      return res.status(404).json({ error: 'Datei nicht verfügbar' });
-    }
-
-    const mediaPath = service.getMediaFilePath(share.file_path);
-
-    try {
-      const stat = await fsPromises.stat(mediaPath);
-      const fileSize = stat.size;
-
-      if (share.media_type === 'video') {
-        const range = req.headers.range;
-
-        if (range) {
-          const parts = range.replace(/bytes=/, '').split('-');
-          const start = parseInt(parts[0], 10);
-          const end = parts[1] ? parseInt(parts[1], 10) : fileSize - 1;
-          const chunkSize = end - start + 1;
-
-          res.writeHead(206, {
-            'Content-Range': `bytes ${start}-${end}/${fileSize}`,
-            'Accept-Ranges': 'bytes',
-            'Content-Length': chunkSize,
-            'Content-Type': 'video/mp4',
-          });
-
-          const stream = fs.createReadStream(mediaPath, { start, end });
-          return stream.pipe(res);
-        } else {
-          res.writeHead(200, {
-            'Content-Length': fileSize,
-            'Content-Type': 'video/mp4',
-          });
-          return fs.createReadStream(mediaPath).pipe(res);
-        }
-      } else {
-        // Support resized thumbnails via ?w= query parameter
-        const requestedWidth = parseInt(req.query.w as string, 10);
-        if (requestedWidth && requestedWidth > 0 && requestedWidth < 2000) {
-          try {
-            const thumbDir = path.join(path.dirname(mediaPath), 'thumbs');
-            const thumbFilename = `${path.basename(mediaPath, path.extname(mediaPath))}_w${requestedWidth}.webp`;
-            const thumbPath = path.join(thumbDir, thumbFilename);
-
-            // Check if cached thumbnail exists
-            try {
-              const thumbStat = await fsPromises.stat(thumbPath);
-              res.setHeader('Content-Type', 'image/webp');
-              res.setHeader('Content-Length', thumbStat.size);
-              res.setHeader('Cache-Control', 'public, max-age=31536000');
-              return fs.createReadStream(thumbPath).pipe(res);
-            } catch {
-              // Thumbnail doesn't exist yet, generate it
-            }
-
-            await fsPromises.mkdir(thumbDir, { recursive: true });
-            const thumbBuffer = await sharp(mediaPath)
-              .resize({ width: requestedWidth })
-              .webp({ quality: 80 })
-              .toBuffer();
-
-            // Cache to disk asynchronously (don't await)
-            fsPromises.writeFile(thumbPath, thumbBuffer).catch((err) => {
-              log.error('Failed to cache thumbnail:', err);
-            });
-
-            res.setHeader('Content-Type', 'image/webp');
-            res.setHeader('Content-Length', thumbBuffer.length);
-            res.setHeader('Cache-Control', 'public, max-age=31536000');
-            return res.send(thumbBuffer);
-          } catch (resizeErr) {
-            log.error('Thumbnail generation failed, serving original:', resizeErr);
-            // Fall through to serve original
-          }
-        }
-
-        res.setHeader('Content-Type', share.mime_type || 'image/png');
-        res.setHeader('Content-Length', fileSize);
-        res.setHeader('Cache-Control', 'public, max-age=31536000');
-        return fs.createReadStream(mediaPath).pipe(res);
-      }
-    } catch {
-      return res.status(404).json({ error: 'Datei nicht gefunden' });
-    }
-  } catch (error) {
-    log.error('Failed to serve preview:', error);
-    return res.status(500).json({ error: 'Fehler beim Laden der Vorschau' });
-  }
-});
-
-/**
- * GET /:shareToken/download
- * Transfer downloads are public (no auth). Image/video downloads require auth.
- */
-router.get(
-  '/:shareToken/download',
-  async (
-    req: Request<ShareTokenParams>,
-    res: Response,
-    next: express.NextFunction
-  ): Promise<void> => {
-    try {
-      const { shareToken } = req.params;
-      const service = await getSharedMediaService();
-      const share = await service.getShareByToken(shareToken as string);
-
-      if (!share) {
-        res.status(404).json({ success: false, error: 'Geteiltes Medium nicht gefunden' });
-        return;
-      }
-
-      if (share.media_type === 'transfer') {
-        if (share.expires_at && new Date(share.expires_at) < new Date()) {
-          res.status(410).json({ success: false, error: 'Dieser Transfer-Link ist abgelaufen.' });
-          return;
-        }
-
-        // scrypt format: "salt:hash"
-        const password = req.headers['x-transfer-password'] as string;
-        if (share.password_hash) {
-          if (!password) {
-            res.status(401).json({ success: false, error: 'Passwort erforderlich.' });
-            return;
-          }
-          const [salt, storedHash] = share.password_hash.split(':');
-          if (!salt || !storedHash) {
-            res.status(500).json({ success: false, error: 'Ungültiger Passwort-Hash.' });
-            return;
-          }
-          const derivedKey = crypto.scryptSync(password, salt, 64).toString('hex');
-          const passwordValid = crypto.timingSafeEqual(
-            Buffer.from(derivedKey, 'hex'),
-            Buffer.from(storedHash, 'hex')
-          );
-          if (!passwordValid) {
-            res.status(403).json({ success: false, error: 'Falsches Passwort.' });
-            return;
-          }
-        }
-
-        const ipAddress = req.ip || req.socket.remoteAddress || 'unknown';
-        await service.recordDownload(shareToken as string, null, ipAddress, share.id);
-
-        try {
-          const transferParams = {
-            user_id: share.user_id,
-            ...(share.wolke_share_link_id != null
-              ? { wolke_share_link_id: share.wolke_share_link_id }
-              : {}),
-            ...(share.wolke_file_path != null ? { wolke_file_path: share.wolke_file_path } : {}),
-            ...(share.file_name != null ? { file_name: share.file_name } : {}),
-          };
-          const result = await transferService.proxyDownloadWithRecord(transferParams);
-
-          res.setHeader(
-            'Content-Type',
-            result.mimeType || share.mime_type || 'application/octet-stream'
-          );
-          res.setHeader('Content-Length', result.size);
-          setContentDisposition(res, result.fileName);
-
-          log.info(`Transfer download: ${(shareToken as string).substring(0, 8)} (public)`);
-          res.send(result.buffer);
-        } catch (proxyErr) {
-          const err = proxyErr as Error;
-          log.error('Transfer proxy download failed', { error: err.message });
-          if (!res.headersSent) {
-            res.status(502).json({ success: false, error: 'Download von Wolke fehlgeschlagen.' });
-          }
-        }
-        return;
-      }
-
-      // Image/video downloads: require auth — pass share to avoid re-fetch
-      (req as RequestWithShare)._share = share;
-      next();
-    } catch (error) {
-      log.error('Failed to process download:', error);
-      if (!res.headersSent) {
-        res.status(500).json({ success: false, error: 'Fehler beim Download' });
-      }
-    }
-  },
-  requireAuth,
-  async (req: Request<ShareTokenParams> & AuthenticatedRequest, res: Response): Promise<void> => {
-    try {
-      const { shareToken } = req.params;
-      const userId = req.user!.id;
-      const userEmail = req.user!.email || 'authenticated-user';
-
-      const service = await getSharedMediaService();
-      const share =
-        (req as RequestWithShare)._share ?? (await service.getShareByToken(shareToken as string));
-
-      if (!share) {
-        res.status(404).json({ success: false, error: 'Geteiltes Medium nicht gefunden' });
-        return;
-      }
-
-      if (share.status === 'processing') {
-        res.status(202).json({
-          success: false,
-          status: 'processing',
-          error: 'Medium wird noch verarbeitet. Bitte warte einen Moment.',
-        });
-        return;
-      }
-
-      if (share.status === 'failed') {
-        res.status(500).json({ success: false, error: 'Verarbeitung fehlgeschlagen' });
-        return;
-      }
-
-      if (!share.file_path) {
-        res.status(404).json({ success: false, error: 'Datei nicht verfügbar' });
-        return;
-      }
-
-      const ipAddress = req.ip || req.socket.remoteAddress || 'unknown';
-      await service.recordDownload(shareToken as string, userEmail, ipAddress);
-
-      const mediaPath = service.getMediaFilePath(share.file_path);
-
-      try {
-        const stat = await fsPromises.stat(mediaPath);
-        const fileSize = stat.size;
-
-        const sanitizedTitle = (share.title || 'media')
-          .replace(/[^a-zA-Z0-9_-]/g, '_')
-          .substring(0, 50);
-
-        const extension =
-          share.media_type === 'video' ? 'mp4' : share.mime_type === 'image/jpeg' ? 'jpg' : 'png';
-        const filename = `${sanitizedTitle}_gruenerator.${extension}`;
-
-        res.setHeader(
-          'Content-Type',
-          share.mime_type || (share.media_type === 'video' ? 'video/mp4' : 'image/png')
-        );
-        res.setHeader('Content-Length', fileSize);
-        setContentDisposition(res, filename);
-
-        log.info(`Download started: ${shareToken} by user ${userId} (${userEmail})`);
-
-        const fileStream = fs.createReadStream(mediaPath);
-        fileStream.pipe(res);
-
-        fileStream.on('error', (streamError): void => {
-          log.error(`Stream error for ${shareToken}: ${streamError.message}`);
-          if (!res.headersSent) {
-            res.status(500).json({ error: 'Fehler beim Download' });
-          }
-        });
-
-        return;
-      } catch {
-        res.status(404).json({ success: false, error: 'Datei nicht gefunden' });
-        return;
-      }
-    } catch (error) {
-      log.error('Failed to download share:', error);
-      if (!res.headersSent) {
-        res.status(500).json({ success: false, error: 'Fehler beim Download' });
-      }
-      return;
-    }
-  }
-);
-
-router.delete(
-  '/:shareToken',
-  requireAuth,
-  async (req: Request<ShareTokenParams> & AuthenticatedRequest, res: Response) => {
-    try {
-      const userId = req.user!.id;
-      const { shareToken } = req.params;
-
-      const service = await getSharedMediaService();
-      await service.deleteShare(userId, shareToken);
-
-      log.info(`Share deleted: ${shareToken} by user ${userId}`);
-
-      res.json({
-        success: true,
-        message: 'Geteiltes Medium gelöscht',
-      });
-    } catch (error) {
-      log.error('Failed to delete share:', error);
-      if (
-        (error as Error).message.includes('not found') ||
-        (error as Error).message.includes('not owned')
-      ) {
-        res.status(404).json({
-          success: false,
-          error: 'Geteiltes Medium nicht gefunden oder keine Berechtigung',
-        });
-      } else {
-        res.status(500).json({
-          success: false,
-          error: 'Geteiltes Medium konnte nicht gelöscht werden',
-        });
-      }
-    }
-  }
-);
-
-// ============================================================================
-// Template Endpoints
-// ============================================================================
-
-/**
- * POST /:shareToken/save-as-template
- * Convert existing shared media to template
- */
-router.post(
-  '/:shareToken/save-as-template',
-  requireAuth,
-  validateBody(saveAsTemplateSchema),
-  async (req: TypedRequest<SaveAsTemplateBody, ShareTokenParams>, res: Response) => {
-    try {
-      const userId = req.user!.id;
-      const userName = req.user?.display_name || req.user?.email || 'Anonymous';
-      const { shareToken } = req.params;
-      const { title, visibility = 'private' } = req.body;
-
-      const service = await getSharedMediaService();
-
-      await service.markAsTemplate(userId, shareToken, title || 'Template', visibility, userName);
-
-      const templateUrl = `/studio?template=${shareToken}`;
-
-      log.info(
-        `Share ${shareToken} marked as template with visibility: ${visibility} by user ${userId}`
-      );
-
-      return res.json({
-        success: true,
-        templateUrl,
-        shareToken,
-        visibility,
-      });
-    } catch (error) {
-      log.error('Failed to save as template:', error);
-      const errorMessage = (error as Error).message;
-      if (errorMessage.includes('not found')) {
-        return res.status(404).json({
-          success: false,
-          error: 'Share not found',
-        });
-      } else if (errorMessage.includes('Not authorized')) {
-        return res.status(403).json({
-          success: false,
-          error: 'Not authorized to mark this as template',
-        });
-      } else {
-        return res.status(500).json({
-          success: false,
-          error: 'Failed to save as template',
-        });
-      }
-    }
-  }
-);
-
-/**
- * POST /templates/:shareToken/clone
- * Clone template to user's gallery
- */
-router.post(
-  '/templates/:shareToken/clone',
-  requireAuth,
-  async (req: Request<ShareTokenParams> & AuthenticatedRequest, res: Response) => {
-    try {
-      const userId = req.user!.id;
-      const userName = req.user?.display_name || req.user?.email || 'Anonymous';
-      const { shareToken } = req.params;
-
-      const service = await getSharedMediaService();
-      const clonedShare = await service.cloneTemplate(shareToken, userId, userName);
-
-      log.info(`Template ${shareToken} cloned to ${clonedShare.shareToken} by user ${userId}`);
-
-      res.json({
-        success: true,
-        share: clonedShare,
-        message: 'Template successfully cloned',
-      });
-    } catch (error) {
-      log.error('Failed to clone template:', error);
-      const errorMessage = (error as Error).message;
-      if (errorMessage.includes('not found')) {
-        res.status(404).json({
-          success: false,
-          error: 'Template not found',
-        });
-      } else if (errorMessage.includes('not accessible') || errorMessage.includes('private')) {
-        res.status(403).json({
-          success: false,
-          error: 'Template not accessible',
-        });
-      } else {
-        res.status(500).json({
-          success: false,
-          error: 'Failed to clone template',
-        });
-      }
-    }
-  }
-);
-
-/**
- * GET /templates
- * List available templates (user's + public)
- */
-router.get('/templates', requireAuth, async (req: AuthenticatedRequest, res: Response) => {
-  try {
-    const userId = req.user!.id;
-    const { visibility } = req.query;
-
-    const service = await getSharedMediaService();
-    const templates = await service.getTemplates(userId, visibility as string);
-
-    res.json({
-      success: true,
-      templates,
-    });
-  } catch (error) {
-    log.error('Failed to get templates:', error);
-    res.status(500).json({
-      success: false,
-      error: 'Failed to retrieve templates',
-    });
-  }
-});
-
-/**
- * GET /templates/:shareToken
- * Get template details (for clone preparation)
- */
-router.get('/templates/:shareToken', (async (
-  req: Request<ShareTokenParams> & { user?: { id: string } },
-  res: Response
-) => {
-  try {
-    const userId = req.user?.id;
-    const { shareToken } = req.params;
-
-    const service = await getSharedMediaService();
-    const template = await service.getTemplateByToken(shareToken, userId);
-
-    res.json({
-      success: true,
-      template,
-    });
-  } catch (error) {
-    log.error('Failed to get template by token:', error);
-    const errorMessage = (error as Error).message;
-    if (errorMessage.includes('not found')) {
-      res.status(404).json({
-        success: false,
-        error: 'Template not found',
-      });
-    } else if (errorMessage.includes('not accessible') || errorMessage.includes('private')) {
-      res.status(403).json({
-        success: false,
-        error: 'Template not accessible',
-      });
-    } else {
-      res.status(500).json({
-        success: false,
-        error: 'Failed to retrieve template',
-      });
-    }
-  }
-}) as unknown as express.RequestHandler);
-
-// ============================================================================
-// PUSH TO PHONE
-// ============================================================================
-
-/**
- * GET /share/devices
- * Get the current user's registered mobile devices (for push-to-phone UI)
- */
-router.get('/devices', requireAuth, async (req: AuthenticatedRequest, res: Response) => {
-  try {
-    const userId = req.user!.id;
-    const { getUserDevices } = await import('../../services/pushNotificationService.js');
-    const devices = await getUserDevices(userId);
-
-    res.json({ success: true, devices });
-  } catch (error) {
-    log.error('Failed to get devices:', error);
-    res.status(500).json({ success: false, error: 'Failed to get devices' });
-  }
-});
-
-/**
- * POST /share/push-to-phone
- * Send a shared media item to the user's mobile device(s) via push notification
- */
-router.post(
-  '/push-to-phone',
-  requireAuth,
-  validateBody(pushToPhoneSchema),
-  async (req: TypedRequest<PushToPhoneBody>, res: Response) => {
-    try {
-      const userId = req.user!.id;
-      const { shareToken } = req.body;
-
-      // Verify the user owns this share
-      const service = await getSharedMediaService();
-      const share = await service.getShareByToken(shareToken);
-
-      if (!share) {
-        return res.status(404).json({ success: false, error: 'Share not found' });
-      }
-
-      if (share.user_id !== userId) {
-        return res.status(403).json({ success: false, error: 'Not authorized' });
-      }
-
-      const mediaType = share.media_type || 'image';
-      const title = mediaType === 'video' ? 'Neues Video empfangen' : 'Neues Bild empfangen';
-      const body = share.title || 'Von Grünerator gesendet';
-
-      const { sendPushToUser } = await import('../../services/pushNotificationService.js');
-      const pushedToDevices = await sendPushToUser(userId, {
-        title,
-        body,
-        data: {
-          type: 'pushed_content',
-          shareToken,
-          mediaType,
-        },
-      });
-
-      log.info(`Push-to-phone: sent to ${pushedToDevices} device(s)`, { userId, shareToken });
-
-      return res.json({ success: true, pushedToDevices });
-    } catch (error) {
-      log.error('Failed to push to phone:', error);
-      return res.status(500).json({ success: false, error: 'Failed to send to phone' });
-    }
-  }
-);
-
-export default router;
