+++ packages/contracts/src/contracts/index.ts
@@ -11,7 +11,7 @@ export { chatGraphContract } from './chatGraphContract.js';
 export { boardsContract } from './boardsContract.js';
 export { boardCommentsContract } from './boardCommentsContract.js';
 export { publicBoardsContract } from './publicBoardsContract.js';
-export { sharesContract } from './sharesContract.js';
+export { sharesContract, sharesReadContract } from './sharesContract.js';
 export { userProfileContract } from './userProfileContract.js';
 export { notebookContract } from './notebookContract.js';
 export { notebookCollectionsContract } from './notebookCollectionsContract.js';
