+++ packages/contracts/src/index.ts
@@ -28,6 +28,7 @@ export {
   boardCommentsContract,
   publicBoardsContract,
   sharesContract,
+  sharesReadContract,
   userProfileContract,
   notebookContract,
   notebookCollectionsContract,
